token varification services implemented

+++ server/handlers.js
@@ -24,11 +24,14 @@ _is.email = (email = '') => {
 };
 
 _is.str = (str = '', minLength = 1, exactLength = false) => {
-  str = str.trim();
   if (exactLength) {
-    str = str.length === minLength ? str : false;
+    str = str.trim().length === minLength ? str.trim() : false;
   }
-  return typeof(str) === 'string' && str.length >= minLength ? str : false;
+  return typeof(str) === 'string' && str.trim().length >= minLength ? str.trim() : false;
+};
+
+_is.token = (str = '') => {
+  return typeof(str) === 'string' && str.trim().length === config.tokenIDLength ? str.trim() : false;
 };
 
 /****************************************************************
@@ -91,20 +94,25 @@ handlers._users.post = (data, cb) => {
 // Users - get
 // Required data: email
 // Optional data: none
-// @TODO authentication
 handlers._users.get = (data, cb) => {
   const email = _is.email(data.query.email);
+  const token = _is.token(data.headers.token);
 
-  if (email) {
-    _db.read('users', email, (err, resp) => {
-      if (!err && resp) {
-        delete resp.password;
-        cb(200, resp);
+  if (email && token) {
+    handlers._tokens.verifyToken(token, email, (tokenIsValid) => {
+      if (tokenIsValid) {
+        _db.read('users', email, (err, resp) => {
+          if (!err && resp) {
+            delete resp.password;
+            cb(200, resp);
+          } else {
+            cb(404);
+          }
+        });
       } else {
-        cb(404);
+        cb(403, {"Error":"Expired or invalid session."});
       }
     });
-
   } else {
     cb(400, {"Error" : "Missing or invalid fields."});
   }
@@ -119,29 +127,36 @@ handlers._users.put = (data, cb) => {
   const firstName = _is.str(data.payload.firstName);
   const lastName = _is.str(data.payload.lastName);
   const password = _is.str(data.payload.password, config.minPasswordLength);
+  const token = _is.token(data.headers.token);
 
-  if (email && (firstName || lastName || password)) {
-    _db.read('users', email, (err, userData) => {
-      if (!err && userData) {
-        if (firstName) {
-          userData.firstName = firstName;
-        }
-        if (lastName) {
-          userData.lastName = lastName;
-        }
-        if (password) {
-          userData.password = helpers.hash(password);
-        }
-        _db.update('users', email, userData, (err) => {
-          if (!err) {
-            cb(200);
+  if (email && token && (firstName || lastName || password)) {
+    handlers._tokens.verifyToken(token, email, (tokenIsValid) => {
+      if (tokenIsValid) {
+        _db.read('users', email, (err, userData) => {
+          if (!err && userData) {
+            if (firstName) {
+              userData.firstName = firstName;
+            }
+            if (lastName) {
+              userData.lastName = lastName;
+            }
+            if (password) {
+              userData.password = helpers.hash(password);
+            }
+            _db.update('users', email, userData, (err) => {
+              if (!err) {
+                cb(200);
+              } else {
+                cb(500, {'Error' : 'Could not update the user.'});
+              }
+            });
+    
           } else {
-            cb(500, {'Error' : 'Could not update the user.'});
+            cb(404, {'Error': 'The user was not found.'});
           }
         });
-
       } else {
-        cb(404, {'Error': 'The user was not found.'});
+        cb(403, {"Error":"Expired or invalid session."});
       }
     });
   } else {
@@ -153,23 +168,29 @@ handlers._users.put = (data, cb) => {
 // Users - delete
 // Required data: email
 // Optional data: none
-// @TODO authentication
 // @TODO delete any other data associated with the user
 handlers._users.delete = (data, cb) => {
   const email = _is.email(data.query.email);
+  const token = _is.token(data.headers.token);
 
-  if (email) {
-    _db.read('users', email, (err) => {
-      if (!err) {
-        _db.delete('users', email, (err) => {
+  if (email && token) {
+    handlers._tokens.verifyToken(token, email, (tokenIsValid) => {
+      if (tokenIsValid) {
+        _db.read('users', email, (err) => {
           if (!err) {
-            cb(200);
+            _db.delete('users', email, (err) => {
+              if (!err) {
+                cb(200);
+              } else {
+                cb(500, {"Error" : "Could not delete the specified user."});
+              }
+            });
           } else {
-            cb(500, {"Error" : "Could not delete the specified user."});
+            cb(404, {"Error" : "Could not find the specified user."});
           }
         });
       } else {
-        cb(404, {"Error" : "Could not find the specified user."});
+        cb(403, {"Error":"Expired or invalid session."});
       }
     });
   } else {
@@ -237,7 +258,7 @@ handlers._tokens.post = (data, cb) => {
 // Optional data: none
 // @TODO authentication
 handlers._tokens.get = (data, cb) => {
-  const id = _is.str(String(data.query.id), config.tokenIDLength, true);
+  const id = _is.token(data.query.id);
 
   if (id) {
     _db.read('tokens', id, (err, tokenData) => {
@@ -258,7 +279,7 @@ handlers._tokens.get = (data, cb) => {
 // Optional data: none
 // @TODO authentication
 handlers._tokens.put = (data, cb) => {
-  const id = _is.str(data.payload.id, config.tokenIDLength, true);
+  const id = _is.token(data.payload.id);
   const extend = typeof(data.payload.extend) === 'boolean' && data.payload.extend === true ? true : false;
 
   if (extend) {
@@ -293,7 +314,7 @@ handlers._tokens.put = (data, cb) => {
 // Optional data: none
 // @TODO authentication
 handlers._tokens.delete = (data, cb) => {
-  const id = _is.str(data.query.id, config.tokenIDLength, true);
+  const id = _is.token(data.query.id);
 
   if (id) {
     _db.read('tokens', id, (err) => {
@@ -311,6 +332,28 @@ handlers._tokens.delete = (data, cb) => {
     });
   } else {
     cb(400, {"Error" : "Missing or invalid fields."});
+  }
+};
+
+// Tokens - session validation
+// Required data: id, email, cb
+// Optional data: none
+handlers._tokens.verifyToken = (id, email, cb) => {
+  id = _is.token(id);
+  email = _is.email(email);
+
+  if (id && email) {
+    _db.read('tokens', id, (err, tokenData) => {
+      if (!err && tokenData) {
+        if (tokenData.expires > Date.now() && tokenData.email === email) {
+          cb(true);
+        } else {
+          cb(false);
+        }
+      } else {
+        cb(false);
+      }
+    });
   }
 };
